ПогодаБот/requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
python-dotenv>=1.0.0
tzdata>=2024.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
